Move TestTaskExecutor to executor package
